Moved files into package - Moved visualizer and preprocessing into pypeline directory - Added support for using template files that are no longer in the same directory
